code fonctionnel
ecriture sd + envoie donne cloudia :)

+++ Classe_Capteurs/Classe_date/classe_date.h
@@ -4,6 +4,7 @@
 #include <inttypes.h>
 #include "Arduino.h"
 #include "TimerOne.h"
+
 #ifndef CLASSE_DATE_H
 #define CLASSE_DATE_H
 
@@ -17,7 +18,7 @@ public:
         unsigned char d = 27;
     };
 
-    int dOffset = 0;
+    long dOffset = 0;
 
     struct sTemp
     {
